Added VN100, 200, 300 support
VectorNav sponsored a rewrite of the driver to support all three
of thier devices. This essentially lead to a rewrite of everything,
so applogies to anyone that was using the old node, you may have a
bit of work to do, but I think you'll agree that it's for the better.

-DAW