Update method for reading data and add odometry (#23)
* Update method for reading data and add odometry

Changed the way data is read. Instead of reading from serial data is read using the built-in CompositeData methods.
Added odomtery messages. After getting a lock on INS, the initial position is recorded and all future postions are published relative to the initial position.

* Add conditional Odom publishing and Odom reset service

Odom will only be published if there is a subscriber to the Odom topic.
Initial position can now be reset to the current location using a ros service call.